requirements.txt
opencv-python==4.10.0
numpy==1.26.4
pandas==2.2.3
Pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
scikit-learn==1.2.2